refactor sagemakerci.cli and apply isort
