+++ cdk-project/lib/images/codebuild-image/python/setup.py
@@ -1,3 +1,3 @@
-from setuptools import setup, find_packages
+from setuptools import find_packages, setup
 
 setup(packages=find_packages("src"), package_dir={"": "src"})
