+++ cdk-project/lib/images/codebuild-image/python/src/sagemakerci/changelog.py
@@ -4,7 +4,6 @@ import logging
 import os
 import re
 
-
 logger = logging.getLogger(__name__)
