+++ cdk-project/lib/images/codebuild-image/python/src/sagemakerci/cli/pr_notebook_filenames.py
@@ -3,9 +3,8 @@ import argparse
 import os
 import sys
 
-from sagemakerci.git import Git
-
 from github import Github
+from sagemakerci.git import Git
 
 
 def parse_args(args):
