+++ cdk-project/lib/images/codebuild-image/python/src/sagemakerci/git.py
@@ -4,7 +4,6 @@ import subprocess
 import urllib
 
 import boto3
-
 import sagemakerci.common
